1. fixed the chapter effects 2. added printing projectiles 3. fixed the cheat spawn zombie

+++ src/com/ussr/pvz/model/engine/modifiers/ModifiableStat.java
@@ -37,7 +37,7 @@ public class ModifiableStat {
         return finalValue;
     }
 
-    public void setBaseValue(int baseValue) {
+    public void setBaseValue(float baseValue) {
         this.baseValue = baseValue;
     }
 }
